Update all broken references due to moving files
Since modules were moved, all references to these modules need to be updated
