+++ gbasis/integrals/electron_repulsion.py
@@ -1,11 +1,11 @@
 """Electron-electron repulsion integral."""
-from gbasis._two_elec_int import (
+from gbasis.base_four_symm import BaseFourIndexSymmetric
+from gbasis.contractions import GeneralizedContractionShell
+from gbasis.integrals._two_elec_int import (
     _compute_two_elec_integrals,
     _compute_two_elec_integrals_angmom_zero,
 )
-from gbasis.base_four_symm import BaseFourIndexSymmetric
-from gbasis.contractions import GeneralizedContractionShell
-from gbasis.point_charge import PointChargeIntegral
+from gbasis.integrals.point_charge import PointChargeIntegral
 import numpy as np
