+++ tests/test_angular_momentum.py
@@ -1,8 +1,10 @@
-"""Test gbasis.angular_momentum."""
-from gbasis._diff_operator_int import _compute_differential_operator_integrals_intermediate
-from gbasis._moment_int import _compute_multipole_moment_integrals_intermediate
-from gbasis.angular_momentum import angular_momentum_integral, AngularMomentumIntegral
+"""Test gbasis.integrals.angular_momentum."""
 from gbasis.contractions import GeneralizedContractionShell
+from gbasis.integrals._diff_operator_int import (
+    _compute_differential_operator_integrals_intermediate,
+)
+from gbasis.integrals._moment_int import _compute_multipole_moment_integrals_intermediate
+from gbasis.integrals.angular_momentum import angular_momentum_integral, AngularMomentumIntegral
 from gbasis.parsers import make_contractions, parse_nwchem
 import numpy as np
 import pytest
@@ -10,7 +12,7 @@ from utils import find_datafile
 
 
 def test_angular_momentum_construct_array_contraction():
-    """Test gbasis.angular_momentum.angular_momentumIntegral.construct_array_contraction."""
+    """Test integrals.angular_momentum.angular_momentumIntegral.construct_array_contraction."""
     test_one = GeneralizedContractionShell(
         1, np.array([0.5, 1, 1.5]), np.array([1.0, 2.0]), np.array([0.1, 0.01])
     )
@@ -365,7 +367,7 @@ def test_angular_momentum_construct_array_contraction():
 
 
 def test_angular_momentum_integral_cartesian():
-    """Test gbasis.angular_momentum.angular_momentum_integral_cartesian."""
+    """Test gbasis.integrals.angular_momentum.angular_momentum_integral_cartesian."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
     basis = make_contractions(basis_dict, ["Kr"], np.array([[0, 0, 0]]))
     angular_momentum_integral_obj = AngularMomentumIntegral(basis)
@@ -376,7 +378,7 @@ def test_angular_momentum_integral_cartesian():
 
 
 def test_angular_momentum_integral_spherical():
-    """Test gbasis.angular_momentum.angular_momentum_integral_spherical."""
+    """Test gbasis.integrals.angular_momentum.angular_momentum_integral_spherical."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
 
     basis = make_contractions(basis_dict, ["Kr"], np.array([[0, 0, 0]]))
@@ -388,7 +390,7 @@ def test_angular_momentum_integral_spherical():
 
 
 def test_angular_momentum_integral_mix():
-    """Test gbasis.angular_momentum.angular_momentum_integral_mix."""
+    """Test gbasis.integrals.angular_momentum.angular_momentum_integral_mix."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
 
     basis = make_contractions(basis_dict, ["Kr"], np.array([[0, 0, 0]]))
@@ -400,7 +402,7 @@ def test_angular_momentum_integral_mix():
 
 
 def test_angular_momentum_integral_lincomb():
-    """Test gbasis.angular_momentum.angular_momentum_integral_lincomb."""
+    """Test gbasis.integrals.angular_momentum.angular_momentum_integral_lincomb."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
     basis = make_contractions(basis_dict, ["Kr"], np.array([[0, 0, 0]]))
     angular_momentum_integral_obj = AngularMomentumIntegral(basis)
