+++ tests/test_one_elec_int.py
@@ -1,6 +1,6 @@
-"""Test gbasis._one_elec_int."""
-from gbasis._one_elec_int import _compute_one_elec_integrals
+"""Test gbasis.integrals._one_elec_int."""
 from gbasis.contractions import GeneralizedContractionShell
+from gbasis.integrals._one_elec_int import _compute_one_elec_integrals
 import numpy as np
 import pytest
 from scipy.special import hyp1f1
